TU de VilleJPADataAccessService en utilisant les mocks: Mockito

+++ src/main/java/com/amdiatou/atiservices/ville/VilleService.java
@@ -71,6 +71,9 @@ public class VilleService {
     }
 
     public void deleteVille(Long id) {
+        if(!villeDao.existsById(id)){
+            throw new ResourceNotFound("La ville avec l'id [%s] n'existe pas".formatted(id));
+        }
         villeDao.deleteVilleById(id);
     }
 }
